requirements.txt
dj-database-url==0.3.0
-e git+https://github.com/django/django.git@449d1effb81152e54f482784cf7febe965007096#egg=Django-dev
django-extensions==1.5.1
-e git+https://github.com/rfadams/django-pgjsonb.git@83306700865ede4c508afe26e263a1268bc68819#egg=django_pgjsonb-master
django-sslify==0.2.5
django-storages==1.1.8
djangorestframework==3.1.0
factory-boy==2.4.1
gunicorn==19.3.0
linecache2==1.0.0
Markdown==2.6.1
psycopg2==2.6
Pygments==2.0.2
pytz==2014.10
PyYAML==3.11
requests==2.6.0
six==1.9.0
South==1.0.2
traceback2==1.4.0
unittest2==1.0.1
yolk==0.4.3
